feat(pages): set a page as an api homepage
fix gravitee-io/issues#467
