+++ gravitee-management-api-service/src/test/java/io/gravitee/management/service/PageServiceTest_Delete.java
@@ -1,0 +1,61 @@
+/**
+ * Copyright (C) 2015 The Gravitee team (http://gravitee.io)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *         http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package io.gravitee.management.service;
+
+import io.gravitee.management.service.exceptions.TechnicalManagementException;
+import io.gravitee.management.service.impl.PageServiceImpl;
+import io.gravitee.repository.exceptions.TechnicalException;
+import io.gravitee.repository.management.api.PageRepository;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.runners.MockitoJUnitRunner;
+
+import static org.mockito.Mockito.doThrow;
+import static org.mockito.Mockito.verify;
+
+/**
+ * @author Azize ELAMRANI (azize.elamrani at graviteesource.com)
+ * @author Nicolas GERAUD (nicolas.geraud at graviteesource.com)
+ * @author GraviteeSource Team
+ */
+@RunWith(MockitoJUnitRunner.class)
+public class PageServiceTest_Delete {
+
+    private static final String PAGE_ID = "ba01aef0-e3da-4499-81ae-f0e3daa4995a";
+
+    @InjectMocks
+    private PageServiceImpl pageService = new PageServiceImpl();
+
+    @Mock
+    private PageRepository pageRepository;
+
+    @Test
+    public void shouldDeletePage() throws TechnicalException {
+        pageService.delete(PAGE_ID);
+
+        verify(pageRepository).delete(PAGE_ID);
+    }
+
+    @Test(expected = TechnicalManagementException.class)
+    public void shouldNotDeletePageBecauseTechnicalException() throws TechnicalException {
+        doThrow(TechnicalException.class).when(pageRepository).delete(PAGE_ID);
+
+        pageService.delete(PAGE_ID);
+    }
+
+}
