+++ gravitee-management-api-service/src/test/resources/io/gravitee/management/service/export-convertAsJsonForExportWithouGroup.json
@@ -12,7 +12,8 @@
     "type" : "MARKDOWN",
     "content" : "Read the doc",
     "order" : 1,
-    "published" : false
+    "published" : false,
+    "homepage" : false
   } ],
   "plans" : [ {
     "id" : "plan-id",
