+++ gravitee-management-api-service/src/test/resources/io/gravitee/management/service/export-convertAsJsonForExportWithoutPlans.json
@@ -12,6 +12,7 @@
     "type" : "MARKDOWN",
     "content" : "Read the doc",
     "order" : 1,
-    "published" : false
+    "published" : false,
+    "homepage" : false
   } ]
 }
