+++ pom.xml
@@ -37,7 +37,7 @@
         <gravitee-definition.version>1.3.0</gravitee-definition.version>
         <gravitee-common.version>1.2.0</gravitee-common.version>
         <gravitee-plugin.version>1.0.0</gravitee-plugin.version>
-        <gravitee-repository.version>1.4.0</gravitee-repository.version>
+        <gravitee-repository.version>1.5.0-SNAPSHOT</gravitee-repository.version>
         <gravitee-gateway-api.version>1.3.0</gravitee-gateway-api.version>
         <gravitee-fetcher-api.version>1.0.0</gravitee-fetcher-api.version>
